Move updating application property file related code from appdetect.go to app_init.go.
